+++ src/api/model/mod.rs
@@ -5,3 +5,4 @@ pub mod xtream;
 pub mod app_state;
 pub mod shared_stream;
 pub mod buffered_stream;
+pub mod persist_pipe_stream;